requirements.txt
asgiref==3.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deap==1.4.3
dill==0.4.0
Django==5.2.10
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.20.2
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multiprocess==0.70.18
mysql-connector-python==9.5.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
pyclipper==1.4.0
pyparsing==3.3.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
sklearn-genetic==0.6.0
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.246
ultralytics-thop==2.0.18
urllib3==2.6.2
Werkzeug==3.1.4
